refactor: delete auth directory.
